+++ src/Facades/Checkout.php
@@ -12,7 +12,27 @@
 namespace Vanilo\Checkout\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Vanilo\Checkout\Contracts\CheckoutState;
+use Vanilo\Contracts\Address;
+use Vanilo\Contracts\Billpayer;
+use Vanilo\Contracts\CheckoutSubject;
 
+/**
+ * @method static getCart(): ?CheckoutSubject
+ * @method static setCart(CheckoutSubject $cart)
+ * @method static getState(): CheckoutState
+ * @method static setState(CheckoutState|string $state)
+ * @method static getBillpayer(): Billpayer
+ * @method static setBillpayer(Billpayer $billpayer)
+ * @method static getShippingAddress(): Address
+ * @method static setShippingAddress(Address $address)
+ * @method static setCustomAttribute(string $key, $value): void
+ * @method static getCustomAttribute(string $key)
+ * @method static setCustomData(array $data): void
+ * @method static getCustomData(): array
+ * @method static update(array $data);
+ * @method static total(): float
+ */
 class Checkout extends Facade
 {
     /**